improvements

+++ src/main/java/com/example/controller/InventoryController.java
@@ -11,6 +11,7 @@ import io.micronaut.http.MediaType;
 import io.micronaut.http.annotation.*;
 import jakarta.inject.Inject;
 import jakarta.inject.Singleton;
+
 import java.util.ArrayList;
 import java.util.List;
 
@@ -23,6 +24,7 @@ import java.util.List;
 @Controller("/inventory")
 public class InventoryController {
     private final InventoryFacade inventoryFacade;
+    final int MIN_STOCK = 5;
 
     @Inject
     public InventoryController(InventoryFacade inventoryFacade) {
@@ -31,9 +33,16 @@ public class InventoryController {
 
     @Get("/{id}")
     @Produces(MediaType.APPLICATION_JSON)
-    public HttpResponse<Item> getItem(@PathVariable String id) {
+    public HttpResponse<Item> getItem(@Header String authorization, @PathVariable String id) {
+        if (authorization == null || !authorization.startsWith("Basic adminInventory")) {
+            return HttpResponse.status(HttpStatus.UNAUTHORIZED, "Unauthorized");
+        }
+
         try {
             Item item = inventoryFacade.getItem(id);
+            if (item == null) {
+                return HttpResponse.status(HttpStatus.NOT_FOUND, "Item not found");
+            }
             return HttpResponse.ok(item);
         } catch (Exception e) {
             e.printStackTrace();
@@ -43,7 +52,11 @@ public class InventoryController {
 
     @Get
     @Produces(MediaType.APPLICATION_JSON)
-    public HttpResponse<ArrayList<Item>> getItems() {
+    public HttpResponse<ArrayList<Item>> getItems(@Header String authorization) {
+        if (authorization == null || !authorization.startsWith("Basic adminInventory")) {
+            return HttpResponse.status(HttpStatus.UNAUTHORIZED, "Unauthorized");
+        }
+
         try {
             ArrayList<Item> items = inventoryFacade.getItems();
             return HttpResponse.ok(items);
@@ -55,7 +68,11 @@ public class InventoryController {
 
     @Get("/export")
     @Produces(MediaType.APPLICATION_XML)
-    public HttpResponse<String> getItemsXML() {
+    public HttpResponse<String> getItemsXML(@Header String authorization) {
+        if (authorization == null || !authorization.startsWith("Basic adminInventory")) {
+            return HttpResponse.status(HttpStatus.UNAUTHORIZED, "Unauthorized");
+        }
+
         try {
             List<Item> items = inventoryFacade.getItems();
             String xml = toXml(items);
@@ -66,7 +83,7 @@ public class InventoryController {
         }
     }
 
-    public String toXml(List<Item> items) throws JsonProcessingException {
+    private String toXml(List<Item> items) throws JsonProcessingException {
         XmlMapper xmlMapper = new XmlMapper();
         return xmlMapper.writeValueAsString(items);
     }
@@ -74,7 +91,16 @@ public class InventoryController {
     @Post
     @Produces(MediaType.APPLICATION_JSON)
     @Consumes(MediaType.APPLICATION_JSON)
-    public HttpResponse<Item> addItem(@Body ItemRequest req) {
+    public HttpResponse<Item> addItem(@Header String authorization, @Body ItemRequest req) {
+        if (authorization == null || !authorization.startsWith("Basic adminInventory")) {
+            return HttpResponse.status(HttpStatus.UNAUTHORIZED, "Unauthorized");
+        }
+
+        // check quantity is not less than the minimum reorder point
+        if (req.getQuantity() <= MIN_STOCK) {
+            return HttpResponse.status(HttpStatus.BAD_REQUEST, "Quantity must be greater than 5");
+        }
+
         try {
             Item item = inventoryFacade.addItem(req.getName(), req.getQuantity(), req.getSupplier(), req.getUnitCost(), req.getMarketingDescription());
             return HttpResponse.ok(item);
@@ -87,13 +113,22 @@ public class InventoryController {
     @Put()
     @Produces(MediaType.APPLICATION_JSON)
     @Consumes(MediaType.APPLICATION_JSON)
-    public HttpResponse<Item> updateItem(@Body Item item) {
+    public HttpResponse<Item> updateItem(@Header String authorization, @Body Item item) {
+        if (authorization == null || !authorization.startsWith("Basic adminInventory")) {
+            return HttpResponse.status(HttpStatus.UNAUTHORIZED, "Unauthorized");
+        }
+
        try {
            boolean updated = inventoryFacade.updateItem(item);
+           // if item quantity is negative, return bad request
+              if (item.getQuantity() < 0) {
+                return HttpResponse.status(HttpStatus.BAD_REQUEST, "Quantity must be greater than 0");
+              }
+
            if (updated) {
                return HttpResponse.ok(item);
            } else {
-               return HttpResponse.status(HttpStatus.INTERNAL_SERVER_ERROR, "Failed to update item");
+               return HttpResponse.status(HttpStatus.NOT_FOUND, "Item not found");
            }
        } catch (Exception e) {
            e.printStackTrace();
@@ -104,13 +139,17 @@ public class InventoryController {
     @Delete("/{id}")
     @Produces(MediaType.APPLICATION_JSON)
     @Consumes(MediaType.APPLICATION_JSON)
-    public HttpResponse<String> deleteItem(@PathVariable String id) {
+    public HttpResponse<String> deleteItem(@Header String authorization, @PathVariable String id) {
+        if (authorization == null || !authorization.startsWith("Basic adminInventory")) {
+            return HttpResponse.status(HttpStatus.UNAUTHORIZED, "Unauthorized");
+        }
+
         try {
             boolean deleted = inventoryFacade.deleteItem(id);
             if (deleted) {
                 return HttpResponse.ok("Item deleted");
             } else {
-                return HttpResponse.status(HttpStatus.INTERNAL_SERVER_ERROR, "Failed to delete item");
+                return HttpResponse.status(HttpStatus.NOT_FOUND, "Item not found");
             }
         } catch (Exception e) {
             e.printStackTrace();
